refactor(host): introduce `providers` module
