+++ crates/host/src/wasmbus/mod.rs
@@ -1,7 +1,5 @@
 #![allow(clippy::type_complexity)]
 
-use core::net::SocketAddr;
-
 use std::collections::btree_map::Entry as BTreeMapEntry;
 use std::collections::hash_map::{self, Entry};
 use std::collections::{BTreeMap, HashMap};
@@ -27,20 +25,15 @@ use cloudevents::{EventBuilder, EventBuilderV10};
 use futures::future::Either;
 use futures::stream::{AbortHandle, Abortable, SelectAll};
 use futures::{join, stream, try_join, Stream, StreamExt, TryFutureExt, TryStreamExt};
-use http::header::HOST;
-use http::uri::Scheme;
-use http::Uri;
-use http_body_util::BodyExt as _;
-use hyper_util::rt::{TokioExecutor, TokioIo};
 use nkeys::{KeyPair, KeyPairType, XKey};
 use secrecy::Secret;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 use tokio::io::{AsyncWrite, AsyncWriteExt};
-use tokio::sync::{broadcast, mpsc, watch, Mutex, RwLock, Semaphore};
+use tokio::sync::{broadcast, mpsc, watch, RwLock, Semaphore};
 use tokio::task::{JoinHandle, JoinSet};
 use tokio::time::{interval_at, Instant};
-use tokio::{fs, process, select, spawn};
+use tokio::{process, select, spawn};
 use tokio_stream::wrappers::IntervalStream;
 use tracing::{debug, error, info, instrument, trace, warn, Instrument as _};
 use uuid::Uuid;
@@ -55,13 +48,7 @@ use wasmcloud_core::{
     provider_config_update_subject, ComponentId, HealthCheckResponse, HostData,
     InterfaceLinkDefinition, OtelConfig, CTL_API_VERSION_1,
 };
-use wasmcloud_provider_messaging_nats::add_tls_ca;
-use wasmcloud_provider_sdk::provider::{
-    handle_provider_commands, receive_link_for_provider, ProviderCommandReceivers,
-};
-use wasmcloud_provider_sdk::{LinkConfig, LinkDeleteInfo, ProviderConnection};
 use wasmcloud_runtime::capability::secrets::store::SecretValue;
-use wasmcloud_runtime::capability::wrpc;
 use wasmcloud_runtime::component::WrpcServeEvent;
 use wasmcloud_runtime::Runtime;
 use wasmcloud_secrets_types::SECRET_PREFIX;
@@ -76,6 +63,7 @@ use crate::{
 
 mod event;
 mod handler;
+mod providers;
 
 pub mod config;
 /// wasmCloud host configuration
@@ -2396,481 +2384,6 @@ impl Host {
     }
 
     #[instrument(level = "debug", skip_all)]
-    async fn start_http_provider(
-        &self,
-        tasks: &mut JoinSet<()>,
-        link_definitions: impl IntoIterator<Item = InterfaceLinkDefinition>,
-        provider_xkey: XKey,
-        host_config: HashMap<String, String>,
-        provider_id: &str,
-        host_id: &str,
-    ) -> anyhow::Result<()> {
-        use wasmcloud_provider_http_server::{
-            default_listen_address, load_settings, ServiceSettings,
-        };
-        use wrpc_interface_http::ServeIncomingHandlerWasmtime as _;
-
-        struct Provider {
-            address: SocketAddr,
-            components: Arc<RwLock<HashMap<String, Arc<Component>>>>,
-            links: Mutex<HashMap<Arc<str>, HashMap<Box<str>, JoinSet<()>>>>,
-            lattice_id: Arc<str>,
-            host_id: Arc<str>,
-        }
-
-        impl wasmcloud_provider_sdk::Provider for Provider {
-            #[instrument(level = "debug", skip_all)]
-            async fn receive_link_config_as_source(
-                &self,
-                LinkConfig {
-                    target_id,
-                    config,
-                    link_name,
-                    ..
-                }: LinkConfig<'_>,
-            ) -> anyhow::Result<()> {
-                let ServiceSettings { address, .. } =
-                    load_settings(Some(self.address), config).context("failed to load settings")?;
-
-                let socket = match &address {
-                    SocketAddr::V4(_) => tokio::net::TcpSocket::new_v4()?,
-                    SocketAddr::V6(_) => tokio::net::TcpSocket::new_v6()?,
-                };
-                socket.set_reuseaddr(!cfg!(windows))?;
-                socket.set_nodelay(true)?;
-                socket.bind(address)?;
-                let listener = socket.listen(8196)?;
-
-                let mut tasks = JoinSet::new();
-                let target_id: Arc<str> = Arc::from(target_id);
-                let svc = hyper::service::service_fn({
-                    let target_id = Arc::clone(&target_id);
-                    let lattice_id = Arc::clone(&self.lattice_id);
-                    let host_id = Arc::clone(&self.host_id);
-                    let components = Arc::clone(&self.components);
-                    move |req: hyper::Request<hyper::body::Incoming>| {
-                        let target_id = Arc::clone(&target_id);
-                        let lattice_id = Arc::clone(&lattice_id);
-                        let host_id = Arc::clone(&host_id);
-                        let components = Arc::clone(&components);
-                        async move {
-                            let component = {
-                                let components = components.read().await;
-                                let component = components
-                                    .get(target_id.as_ref())
-                                    .context("linked component not found")?;
-                                Arc::clone(component)
-                            };
-                            let (
-                                http::request::Parts {
-                                    method,
-                                    uri,
-                                    headers,
-                                    ..
-                                },
-                                body,
-                            ) = req.into_parts();
-                            let http::uri::Parts {
-                                scheme,
-                                authority,
-                                path_and_query,
-                                ..
-                            } = uri.into_parts();
-                            let mut uri = Uri::builder().scheme(scheme.unwrap_or(Scheme::HTTP));
-                            if let Some(authority) = authority {
-                                uri = uri.authority(authority);
-                            } else if let Some(authority) = headers.get("X-Forwarded-Host") {
-                                uri = uri.authority(authority.as_bytes());
-                            } else if let Some(authority) = headers.get(HOST) {
-                                uri = uri.authority(authority.as_bytes());
-                            }
-                            if let Some(path_and_query) = path_and_query {
-                                uri = uri.path_and_query(path_and_query)
-                            };
-                            let uri = uri.build().context("invalid URI")?;
-                            let mut req = http::Request::builder().method(method);
-                            *req.headers_mut().expect("headers missing") = headers;
-
-                            let req = req
-                                .uri(uri)
-                                .body(
-                                    body.map_err(wasmtime_wasi_http::hyper_response_error)
-                                        .boxed(),
-                                )
-                                .context("invalid request")?;
-                            let res = component
-                                .instantiate(
-                                    component.handler.copy_for_new(),
-                                    component.events.clone(),
-                                )
-                                .handle(
-                                    (
-                                        Instant::now(),
-                                        vec![
-                                            KeyValue::new(
-                                                "component.ref",
-                                                Arc::clone(&component.image_reference),
-                                            ),
-                                            KeyValue::new("lattice", Arc::clone(&lattice_id)),
-                                            KeyValue::new("host", Arc::clone(&host_id)),
-                                        ],
-                                    ),
-                                    req,
-                                )
-                                .await?;
-                            let res = res?;
-                            anyhow::Ok(res)
-                        }
-                    }
-                });
-
-                let srv = hyper_util::server::conn::auto::Builder::new(TokioExecutor::new());
-                let srv = Arc::new(srv);
-
-                tasks.spawn(async move {
-                    loop {
-                        let stream = match listener.accept().await {
-                            Ok((stream, _)) => stream,
-                            Err(err) => {
-                                error!(?err, "failed to accept HTTP server connection");
-                                continue;
-                            }
-                        };
-                        let svc = svc.clone();
-                        tokio::spawn({
-                            let srv = Arc::clone(&srv);
-                            let svc = svc.clone();
-                            async move {
-                                if let Err(err) =
-                                    srv.serve_connection(TokioIo::new(stream), svc).await
-                                {
-                                    error!(?err, "failed to serve connection");
-                                }
-                            }
-                        });
-                    }
-                });
-                self.links
-                    .lock()
-                    .await
-                    .entry(target_id)
-                    .or_default()
-                    .insert(link_name.into(), tasks);
-                Ok(())
-            }
-
-            #[instrument(level = "debug", skip_all)]
-            async fn delete_link_as_source(&self, info: impl LinkDeleteInfo) -> anyhow::Result<()> {
-                let target_id = info.get_target_id();
-                let link_name = info.get_link_name();
-                self.links
-                    .lock()
-                    .await
-                    .get_mut(target_id)
-                    .map(|links| links.remove(link_name));
-                Ok(())
-            }
-        }
-
-        match host_config.get("routing_mode").map(String::as_str) {
-            // Run provider in address mode by default
-            Some("address") | None => {}
-            // Run provider in path mode
-            Some("path") => bail!("path mode not supported by builtin yet"),
-            Some(other) => bail!("unknown routing_mode: {other}"),
-        };
-        let default_address = host_config
-            .get("default_address")
-            .map(|s| SocketAddr::from_str(s))
-            .transpose()
-            .context("failed to parse default_address")?
-            .unwrap_or_else(default_listen_address);
-
-        let (quit_tx, quit_rx) = broadcast::channel(1);
-        let commands = ProviderCommandReceivers::new(
-            Arc::clone(&self.rpc_nats),
-            &quit_tx,
-            &self.host_config.lattice,
-            provider_id,
-            provider_id,
-            host_id,
-        )
-        .await?;
-        let conn = ProviderConnection::new(
-            Arc::clone(&self.rpc_nats),
-            Arc::from(provider_id),
-            Arc::clone(&self.host_config.lattice),
-            host_id.to_string(),
-            host_config,
-            provider_xkey,
-            Arc::clone(&self.secrets_xkey),
-        )
-        .context("failed to establish provider connection")?;
-        let provider = Provider {
-            address: default_address,
-            components: Arc::clone(&self.components),
-            links: Mutex::default(),
-            host_id: Arc::from(host_id),
-            lattice_id: Arc::clone(&self.host_config.lattice),
-        };
-        for ld in link_definitions {
-            if let Err(e) = receive_link_for_provider(&provider, &conn, ld).await {
-                error!(
-                    error = %e,
-                    "failed to initialize link during provider startup",
-                );
-            }
-        }
-        tasks.spawn(async move {
-            handle_provider_commands(provider, &conn, quit_rx, quit_tx, commands).await
-        });
-        Ok(())
-    }
-
-    #[instrument(level = "debug", skip_all)]
-    async fn start_messaging_nats_provider(
-        &self,
-        tasks: &mut JoinSet<()>,
-        link_definitions: impl IntoIterator<Item = InterfaceLinkDefinition>,
-        provider_xkey: XKey,
-        host_config: HashMap<String, String>,
-        provider_id: &str,
-        host_id: &str,
-    ) -> anyhow::Result<()> {
-        use wasmcloud_provider_messaging_nats::ConnectionConfig;
-
-        struct Provider {
-            config: ConnectionConfig,
-            components: Arc<RwLock<HashMap<String, Arc<Component>>>>,
-            messaging_links:
-                Arc<RwLock<HashMap<Arc<str>, Arc<RwLock<HashMap<Box<str>, async_nats::Client>>>>>>,
-            subscriptions: Mutex<HashMap<Arc<str>, HashMap<Box<str>, JoinSet<()>>>>,
-            lattice_id: Arc<str>,
-            host_id: Arc<str>,
-        }
-
-        impl Provider {
-            async fn connect(
-                &self,
-                config: &HashMap<String, String>,
-            ) -> anyhow::Result<(async_nats::Client, ConnectionConfig)> {
-                // NOTE: Big part of this is copy-pasted from `provider-messaging-nats`
-                let config = if config.is_empty() {
-                    self.config.clone()
-                } else {
-                    match ConnectionConfig::from_map(config) {
-                        Ok(cc) => self.config.merge(&cc),
-                        Err(err) => {
-                            error!(?err, "failed to build connection configuration");
-                            return Err(anyhow!(err).context("failed to build connection config"));
-                        }
-                    }
-                };
-                let mut opts = match (&config.auth_jwt, &config.auth_seed) {
-                    (Some(jwt), Some(seed)) => {
-                        let seed =
-                            KeyPair::from_seed(seed).context("failed to parse seed key pair")?;
-                        let seed = Arc::new(seed);
-                        async_nats::ConnectOptions::with_jwt(jwt.clone(), move |nonce| {
-                            let seed = seed.clone();
-                            async move { seed.sign(&nonce).map_err(async_nats::AuthError::new) }
-                        })
-                    }
-                    (None, None) => async_nats::ConnectOptions::default(),
-                    _ => bail!("must provide both jwt and seed for jwt authentication"),
-                };
-                if let Some(tls_ca) = &config.tls_ca {
-                    opts = add_tls_ca(tls_ca, opts)?;
-                } else if let Some(tls_ca_file) = &config.tls_ca_file {
-                    let ca = fs::read_to_string(tls_ca_file)
-                        .await
-                        .context("failed to read TLS CA file")?;
-                    opts = add_tls_ca(&ca, opts)?;
-                }
-
-                // Use the first visible cluster_uri
-                let url = config.cluster_uris.first().context("invalid address")?;
-
-                // Override inbox prefix if specified
-                if let Some(ref prefix) = config.custom_inbox_prefix {
-                    opts = opts.custom_inbox_prefix(prefix);
-                }
-                let nats = opts
-                    .name("builtin NATS Messaging Provider")
-                    .connect(url)
-                    .await
-                    .context("failed to connect to NATS")?;
-                Ok((nats, config))
-            }
-        }
-
-        impl wasmcloud_provider_sdk::Provider for Provider {
-            #[instrument(level = "debug", skip_all)]
-            async fn receive_link_config_as_target(
-                &self,
-                LinkConfig {
-                    source_id,
-                    link_name,
-                    config,
-                    ..
-                }: LinkConfig<'_>,
-            ) -> anyhow::Result<()> {
-                let (nats, _) = self.connect(config).await?;
-                let mut links = self.messaging_links.write().await;
-                let mut links = links.entry(source_id.into()).or_default().write().await;
-                links.insert(link_name.into(), nats);
-                Ok(())
-            }
-
-            #[instrument(level = "debug", skip_all)]
-            async fn receive_link_config_as_source(
-                &self,
-                LinkConfig {
-                    target_id,
-                    config,
-                    link_name,
-                    ..
-                }: LinkConfig<'_>,
-            ) -> anyhow::Result<()> {
-                use wrpc::exports::wasmcloud::messaging::handler::Handler as _;
-
-                let (nats, config) = self.connect(config).await?;
-                let mut tasks = JoinSet::new();
-                let target_id: Arc<str> = Arc::from(target_id);
-                for sub in config.subscriptions {
-                    if sub.is_empty() {
-                        continue;
-                    }
-                    let mut sub = if let Some((subject, queue)) = sub.split_once('|') {
-                        nats.queue_subscribe(async_nats::Subject::from(subject), queue.into())
-                            .await
-                    } else {
-                        nats.subscribe(async_nats::Subject::from(sub)).await
-                    }
-                    .context("failed to subscribe")?;
-                    let target_id = Arc::clone(&target_id);
-                    let lattice_id = Arc::clone(&self.lattice_id);
-                    let host_id = Arc::clone(&self.host_id);
-                    let components = Arc::clone(&self.components);
-                    tasks.spawn(async move {
-                        while let Some(async_nats::Message {
-                            subject,
-                            payload,
-                            reply,
-                            ..
-                        }) = sub.next().await
-                        {
-                            let component = {
-                                let components = components.read().await;
-                                let Some(component) = components.get(target_id.as_ref()) else {
-                                    warn!(?target_id, "linked component not found");
-                                    continue;
-                                };
-                                Arc::clone(component)
-                            };
-                            match component
-                                .instantiate(
-                                    component.handler.copy_for_new(),
-                                    component.events.clone(),
-                                )
-                                .handle_message(
-                                    (
-                                        Instant::now(),
-                                        vec![
-                                            KeyValue::new(
-                                                "component.ref",
-                                                Arc::clone(&component.image_reference),
-                                            ),
-                                            KeyValue::new("lattice", Arc::clone(&lattice_id)),
-                                            KeyValue::new("host", Arc::clone(&host_id)),
-                                        ],
-                                    ),
-                                    wrpc::wasmcloud::messaging::types::BrokerMessage {
-                                        subject: subject.into_string(),
-                                        body: payload,
-                                        reply_to: reply.map(async_nats::Subject::into_string),
-                                    },
-                                )
-                                .await
-                            {
-                                Ok(Ok(())) => {}
-                                Ok(Err(err)) => {
-                                    warn!(?err, "component failed to handle message")
-                                }
-                                Err(err) => {
-                                    warn!(?err, "failed to call component")
-                                }
-                            }
-                        }
-                    });
-                }
-                self.subscriptions
-                    .lock()
-                    .await
-                    .entry(target_id)
-                    .or_default()
-                    .insert(link_name.into(), tasks);
-                Ok(())
-            }
-
-            #[instrument(level = "debug", skip_all)]
-            async fn delete_link_as_source(&self, info: impl LinkDeleteInfo) -> anyhow::Result<()> {
-                let target_id = info.get_target_id();
-                let link_name = info.get_link_name();
-                self.subscriptions
-                    .lock()
-                    .await
-                    .get_mut(target_id)
-                    .map(|links| links.remove(link_name));
-                Ok(())
-            }
-        }
-
-        let config = ConnectionConfig::from_map(&host_config).context("failed to parse config")?;
-
-        let (quit_tx, quit_rx) = broadcast::channel(1);
-        let commands = ProviderCommandReceivers::new(
-            Arc::clone(&self.rpc_nats),
-            &quit_tx,
-            &self.host_config.lattice,
-            provider_id,
-            provider_id,
-            host_id,
-        )
-        .await?;
-        let conn = ProviderConnection::new(
-            Arc::clone(&self.rpc_nats),
-            Arc::from(provider_id),
-            Arc::clone(&self.host_config.lattice),
-            host_id.to_string(),
-            host_config,
-            provider_xkey,
-            Arc::clone(&self.secrets_xkey),
-        )
-        .context("failed to establish provider connection")?;
-        let provider = Provider {
-            config,
-            components: Arc::clone(&self.components),
-            messaging_links: Arc::clone(&self.messaging_links),
-            subscriptions: Mutex::default(),
-            host_id: Arc::from(host_id),
-            lattice_id: Arc::clone(&self.host_config.lattice),
-        };
-        for ld in link_definitions {
-            if let Err(e) = receive_link_for_provider(&provider, &conn, ld).await {
-                error!(
-                    error = %e,
-                    "failed to initialize link during provider startup",
-                );
-            }
-        }
-        tasks.spawn(async move {
-            handle_provider_commands(provider, &conn, quit_rx, quit_tx, commands).await
-        });
-        Ok(())
-    }
-
-    #[instrument(level = "debug", skip_all)]
     async fn handle_start_provider_task(
         &self,
         config: &[String],
@@ -3025,7 +2538,7 @@ impl Host {
                 }
                 (None, ResourceRef::Builtin(name)) => match *name {
                     "http-server" => {
-                        self.start_http_provider(
+                        self.start_http_server_provider(
                             &mut tasks,
                             link_definitions,
                             provider_xkey,
