+++ crates/host/src/wasmbus/providers/mod.rs
@@ -1,0 +1,2 @@
+mod http_server;
+mod messaging_nats;
